Use separate Jest preprocessors based on file path
Some Babel plugins are only meant to run on source files, some only on
tests, and so on. We were using a single preprocessor and some
metaprogramming to determine which ones to run per file; I changed it
to use separate preprocessors per category instead.

The practical reason I made this change was so the error code script
does not run on test files.
