+++ scripts/jest/preprocessorForCoffeeScript.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const coffee = require('coffee-script');
+
+module.exports = {
+  process(src, filePath) {
+    return coffee.compile(src, {bare: true});
+  },
+};
